third_party/blink/web_tests/platform/mac/external/wpt/css/css-flexbox/percent-height-flex-items-cross-sizes-with-mutations-expected.txt

This is a testharness.js-based test.
[FAIL] percent-height-flex-items-cross-sizes-with-mutations
  assert_equals: height expected 40 but got 36
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 1
  assert_equals: height expected 40 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 2
  assert_equals: height expected 40 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 3
  assert_equals: height expected 40 but got 36
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 4
  assert_equals: height expected 40 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 5
  assert_equals: height expected 40 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 6
  assert_equals: height expected 30 but got 36
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 7
  assert_equals: height expected 30 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 8
  assert_equals: height expected 10 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 9
  assert_equals: height expected 10 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 10
  assert_equals: height expected 30 but got 36
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 11
  assert_equals: height expected 30 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 12
  assert_equals: height expected 10 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 14
  assert_equals: height expected 100 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 15
  assert_equals: height expected 100 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 16
  assert_equals: height expected 30 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 17
  assert_equals: height expected 30 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 18
  assert_equals: height expected 100 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 19
  assert_equals: height expected 30 but got 18
[FAIL] percent-height-flex-items-cross-sizes-with-mutations 20
  assert_equals: height expected 30 but got 18
Harness: the test ran to completion.
